+++ Store.Client/src/app/app-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { ProductComponent } from '../components/product/product.component';
 import { ProductDetailComponent } from '../components/product-detail/product-detail.component';
-import { ProductListComponent } from '../components/product-list/product-list.component';
+import { ProductListComponent } from './product-list/product-list.component';
 
 const routes: Routes = [
   { path: '', component: ProductListComponent },
