+++ Store.Client/src/app/app.module.ts
@@ -6,7 +6,7 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { ProductComponent } from '../components/product/product.component';
 import { ProductDetailComponent } from '../components/product-detail/product-detail.component';
-import { ProductListComponent } from '../components/product-list/product-list.component';
+import { ProductListComponent } from './product-list/product-list.component';
 
 @NgModule({
   declarations: [
