+++ Store.Client/src/app/models/product.ts
@@ -1,0 +1,8 @@
+export interface Product {
+    idxCode: string;
+    idxCodeAlt: string;
+    name: string | null;
+    date: string | null;
+    quantity: number | null;
+    price: number | null;
+  }
