+++ Store.Client/src/app/product-list/product-list.component.ts
