+++ Store.Client/src/components/product-detail/product-detail.component.html
@@ -1,7 +1,9 @@
-<h1 id="productDetail">Test product detail</h1>
-
-<div class="product-detail-wrapper">
-    <div *ngIf="product" class="product-detail">
-        <div class="test"></div>
+<div *ngIf="productDetails" class="product-card-wrapper">
+    <div class="product-card" [style.backgroundImage]="'url(' + imageUrl + ')'">
+      Product Details
     </div>
-</div>
+    <div class="product-details">
+      <p>IdxCode: {{ productDetails.idxCode }}</p>
+      <p>IdxCodeAlt: {{ productDetails.idxCodeAlt }}</p>
+    </div>
+  </div>
