+++ Store.Client/src/components/product/product.component.html
@@ -8,5 +8,4 @@
     <p>IdxCode: {{ product.idxCode }}</p>
     <p>IdxCodeAlt: {{ product.idxCodeAlt }}</p>
   </div>
-
 </div>
